+++ CHANGELOG.md
@@ -12,12 +12,17 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 
 * renamed ```run_sample.sh``` -> ```run-sample.sh```
-* ```ttlInSeconds``` property is now ```ttl``` and the value associated
+* the ```ttlInSeconds``` property in spider metadata is now ```ttl``` and the value associated
 with the property is now a string instead of an integer - the string
 has the form ```<number><duration>```
 where ```<number>``` is an integer and ```<duration>``` is one
 of ```s```, ```m```, ```h``` or ```d``` representing seconds, minutes, hours
 and days respectively
+* the ```maxCrawlTimeInSeconds``` spider metadata property is now ```maxCrawlTime```
+and is also a string instead of an integer - the string
+has the form ```<number><duration>```
+where ```<number>``` is an integer and ```<duration>``` is one
+of ```s``` or ```m``` representing seconds and minutes respectively
 
 ### Removed